perf: implement sharded recorder (#13724)
dashmap is effectively the same hashmap under lock just with multiple
shards (roughly corresponding to available parallelism) which should
mean that changes to the recorder should rarely contend each other.

It does otherwise complicate the implementation a bit. Especially hard
to reason about were the changes to the methods that formerly drained
the recorder's contents. Turns out that making it take whole `Trie` by
value still works just fine for production code and only required
weirder workarounds in some of the tests.

To be done: performance effect of this PR in isolation.

Related PR: https://github.com/near/nearcore/pull/13702

+++ runtime/runtime/src/tests/apply.rs
@@ -1449,7 +1449,7 @@ fn test_exclude_contract_code_from_witness() {
     // Set the storage proof soft-limit to the size of the contract.
     // Since contract code is not included in the storage proof, both function calls below pass the proof soft-limit.
     let mut runtime_config = RuntimeConfig::test();
-    runtime_config.witness_config.main_storage_proof_size_soft_limit = CONTRACT_SIZE;
+    runtime_config.witness_config.main_storage_proof_size_soft_limit = CONTRACT_SIZE as u64;
     apply_state.config = Arc::new(runtime_config);
 
     let contract_code =
